~inexpdescendantsof ~inancestorssof
