+++ query/src/org/labkey/query/QueryServiceImpl.java
@@ -40,6 +40,8 @@ import org.labkey.api.cache.Cache;
 import org.labkey.api.cache.CacheManager;
 import org.labkey.api.collections.CaseInsensitiveHashMap;
 import org.labkey.api.collections.LabKeyCollectors;
+import org.labkey.api.exp.query.ExpTable;
+import org.labkey.api.query.column.BuiltInColumnTypes;
 import org.labkey.api.query.column.ColumnInfoTransformer;
 import org.labkey.api.data.*;
 import org.labkey.api.data.dialect.SqlDialect;
@@ -54,6 +56,7 @@ import org.labkey.api.module.ModuleResourceCaches;
 import org.labkey.api.module.ResourceRootProvider;
 import org.labkey.api.query.*;
 import org.labkey.api.query.QueryChangeListener.QueryPropertyChange;
+import org.labkey.api.query.snapshot.AbstractTableMethodInfo;
 import org.labkey.api.query.snapshot.QuerySnapshotDefinition;
 import org.labkey.api.resource.Resource;
 import org.labkey.api.security.User;
@@ -97,21 +100,7 @@ import org.labkey.query.persist.LinkedSchemaDef;
 import org.labkey.query.persist.QueryDef;
 import org.labkey.query.persist.QueryManager;
 import org.labkey.query.persist.QuerySnapshotDef;
-import org.labkey.query.sql.Method;
-import org.labkey.query.sql.QDot;
-import org.labkey.query.sql.QExpr;
-import org.labkey.query.sql.QField;
-import org.labkey.query.sql.QIfDefined;
-import org.labkey.query.sql.QInternalExpr;
-import org.labkey.query.sql.QMethodCall;
-import org.labkey.query.sql.QNode;
-import org.labkey.query.sql.QQuery;
-import org.labkey.query.sql.QRowStar;
-import org.labkey.query.sql.QUnion;
-import org.labkey.query.sql.Query;
-import org.labkey.query.sql.QueryTableInfo;
-import org.labkey.query.sql.SqlBuilder;
-import org.labkey.query.sql.SqlParser;
+import org.labkey.query.sql.*;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.web.servlet.mvc.Controller;
 
@@ -138,6 +127,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
@@ -247,7 +237,9 @@ public class QueryServiceImpl implements QueryService
             CompareType.MV_INDICATOR,
             CompareType.NO_MV_INDICATOR,
             CompareType.Q,
-            WHERE
+            WHERE,
+            INDESCENDANTSOF,
+            INANCESTORSOF
     ));
 
     public static final CompareType WHERE = new CompareType("WHERE", "where", "WHERE", true /* dataValueRequired */, "sql", OperatorType.WHERE)
@@ -259,8 +251,50 @@ public class QueryServiceImpl implements QueryService
         }
     };
 
+    public static final CompareType INDESCENDANTSOF = new CompareType("IN DESCENDANTS OF", "inexpdescendantsof", "IN_EXP_DESCENDANTS_OF", true /* dataValueRequired */, "sql", OperatorType.INEXPDESCENDANTSOF)
+    {
+        @Override
+        public WhereClause createFilterClause(@NotNull FieldKey fieldKey, Object value)
+        {
+            return InLineageOfClause.create(false, fieldKey, (String) value);
+        }
+    };
+
+    public static final CompareType INANCESTORSOF = new CompareType("IN ANCESTORS OF", "inexpancestorsof", "IN_EXP_ANCESTORS_OF", true /* dataValueRequired */, "sql", OperatorType.INEXPANCESTORSOF)
+    {
+        @Override
+        public WhereClause createFilterClause(@NotNull FieldKey fieldKey, Object value)
+        {
+            return InLineageOfClause.create(true, fieldKey, (String) value);
+        }
+    };
+
+
+    /*
+     * This is a marker for CompareClauses that can somewhat participate in Query (LabKey SQL) environment.
+     *
+     * At the moment this pretty much means the CompareClause might create sub-select expressions.  In the
+     * future it may be that it resolves columns directly against QueryRelation instead of a TableInfo or other
+     * optimizations/integrations.
+     */
+    public static class QueryCompareClause extends CompareType.CompareClause
+    {
+        protected Query query = null;
+
+        public QueryCompareClause(FieldKey fieldKey, CompareType compareType, Object value)
+        {
+            super(fieldKey, compareType, value);
+        }
+
+        public void setQuery(Query query)
+        {
+            this.query = query;
+        }
+    }
+
+
     // This is defined here due to the usage of QExpr
-    private static class WhereClause extends CompareType.CompareClause
+    private static class WhereClause extends QueryCompareClause
     {
         final QExpr expr;
 
@@ -303,7 +337,7 @@ public class QueryServiceImpl implements QueryService
             return new ArrayList<>(set);
         }
 
-        private void collectKeys(QExpr expr, Set<FieldKey> set)
+        protected void collectKeys(QExpr expr, Set<FieldKey> set)
         {
             if (expr instanceof QQuery || expr instanceof QUnion || expr instanceof QRowStar || expr instanceof QIfDefined)
                 return;
@@ -333,6 +367,14 @@ public class QueryServiceImpl implements QueryService
         @Override
         public SQLFragment toSQLFragment(Map<FieldKey, ? extends ColumnInfo> columnMap, SqlDialect dialect)
         {
+            if (null == query)
+            {
+                Optional<? extends ColumnInfo> c = columnMap.isEmpty() ? Optional.empty() : columnMap.values().stream().findAny();
+                UserSchema s = c.isEmpty() ? null : c.get().getParentTable().getUserSchema();
+                if (null == s)
+                    throw new QueryException("unsupported filter");
+                query = new Query(s);
+            }
             String expression = (String) getParamVals()[0];
             if (isBlank(expression))
                 return new SQLFragment("1=1");
@@ -342,14 +384,32 @@ public class QueryServiceImpl implements QueryService
             if (null == expr || !errors.isEmpty())
                 return new SQLFragment("0=1"); // UNDONE: error reporting
             QExpr bound = bind(expr, columnMap);
-            return bound.getSqlFragment(dialect, null);
+            SQLFragment ret = null;
+            if (null != bound && query.getParseErrors().isEmpty())
+                ret = bound.getSqlFragment(dialect, query);
+            if (!query.getParseErrors().isEmpty() || null==ret)
+            {
+                if (query.getParseErrors().isEmpty())
+                    query.getParseErrors().add(new QueryException("Unexpected error parsing where filter"));
+                throw query.getParseErrors().get(0);
+            }
+            return ret;
         }
 
         private QExpr bind(QExpr expr, Map<FieldKey, ? extends ColumnInfo> columnMap)
         {
-            if (expr instanceof QQuery || expr instanceof QUnion || expr instanceof QRowStar || expr instanceof QIfDefined)
+            if (expr instanceof QUnion || expr instanceof QRowStar || expr instanceof QIfDefined)
             {
-                throw new UnsupportedOperationException("unsupported expression");
+                query.getParseErrors().add(new QueryException("Expression is not supported in where filter"));
+                return null;
+            }
+
+            /* See QuerySelect.declareFields() */
+            if (expr instanceof QQuery qquery)
+            {
+                QueryRelation select = Query.createQueryRelation(query, qquery, true);
+                select.declareFields();
+                return new QQuery(select);
             }
 
             FieldKey key = expr.getFieldKey();
@@ -357,7 +417,10 @@ public class QueryServiceImpl implements QueryService
             {
                 ColumnInfo c = columnMap.get(key);
                 if (null == c)
-                    throw new UnsupportedOperationException("column not found: " + key.toString());
+                {
+                    query.getParseErrors().add(new QueryException("column not found: " + key.toString()));
+                    return null;
+                }
                 return new QColumnInfo(c);
             }
 
@@ -377,9 +440,56 @@ public class QueryServiceImpl implements QueryService
                 else
                     ret.appendChild(bind((QExpr) child, columnMap));
             }
+            if (!query.getParseErrors().isEmpty())
+                return null;
             return ret;
         }
     }
+
+
+    private static final FieldKey expObjectIdFieldKey = new FieldKey(null, BuiltInColumnTypes.EXPOBJECTID_CONCEPT_URI);
+
+
+    /* It would be nice to put with class ChildOfMethod and class ParentOfMethod, but our build dependencies mean this goes here for now */
+    private static class InLineageOfClause extends WhereClause
+    {
+        static InLineageOfClause create(boolean parents, FieldKey fk, String value)
+        {
+            if (null != fk.getParent() || !StringUtils.startsWithIgnoreCase(value, "SELECT "))
+                throw new IllegalArgumentException("Bad filter syntax, value should start with 'SELECT '");
+            return new InLineageOfClause(parents, value);
+        }
+
+        InLineageOfClause(boolean parents, String value)
+        {
+            // I don't have a table info available.  So I can't tell a) if this is an ExpTable b) if
+            // this table Supports expObject() method.  Due to a hack in QuerySeriveImpl.ensureRequiredColumns()
+            // The object id column will be available if that method is supported, but I have to find it later.
+            super("\"" + BuiltInColumnTypes.EXPOBJECTID_CONCEPT_URI + "\" IN " + (parents ? "EXPANCESTORSOF" : "EXPDESCENDANTSOF") + " (" + value + ")");
+        }
+
+        @Override
+        public List<FieldKey> getFieldKeys()
+        {
+            var list = super.getFieldKeys();
+            list.add(expObjectIdFieldKey);
+            return list;
+        }
+
+        @Override
+        public void setSelectColumns(List<ColumnInfo> selectColumns)
+        {
+            super.setSelectColumns(selectColumns);
+        }
+
+        @Override
+        public SQLFragment toSQLFragment(Map<FieldKey, ? extends ColumnInfo> columnMap, SqlDialect dialect)
+        {
+            // Unfortunately
+            return super.toSQLFragment(columnMap, dialect);
+        }
+    }
+
 
     public static class QColumnInfo extends QInternalExpr
     {
@@ -388,6 +498,14 @@ public class QueryServiceImpl implements QueryService
         QColumnInfo(ColumnInfo col)
         {
             _col = col;
+        }
+
+        @Override
+        public ColumnInfo createColumnInfo(SQLTableInfo table, String name, Query query)
+        {
+            var ret = super.createColumnInfo(table, name, query);
+            _col.copyTo((ColumnRenderPropertiesImpl) ret);
+            return ret;
         }
 
         @Override
@@ -1389,7 +1507,7 @@ public class QueryServiceImpl implements QueryService
         return createQuerySnapshotDef(queryDef.getContainer(), queryDef, name);
     }
 
-    private ColumnInfo getColumn(AliasManager manager, TableInfo table, Map<FieldKey, ColumnInfo> columnMap, FieldKey key)
+    public ColumnInfo getColumn(AliasManager manager, TableInfo table, Map<FieldKey, ColumnInfo> columnMap, FieldKey key)
     {
         if (key != null && key.getTable() == null)
         {
@@ -1549,7 +1667,7 @@ public class QueryServiceImpl implements QueryService
 
 
     // mapping may include multiple fieldkeys pointing at same columninfo (see ColumnInfo.resolveColumn());
-    private List<ColumnInfo> ensureRequiredColumns(@NotNull TableInfo table, @NotNull Collection<ColumnInfo> columns, @Nullable Filter filter,
+    public List<ColumnInfo> ensureRequiredColumns(@NotNull TableInfo table, @NotNull Collection<ColumnInfo> columns, @Nullable Filter filter,
                                                    @Nullable Sort sort, @Nullable Set<FieldKey> unresolvedColumns,
                                                    Map<FieldKey, ColumnInfo> columnMap /* IN/OUT */, Set<ColumnInfo> allInvolvedColumns /* IN/OUT */)
     {
@@ -1707,6 +1825,22 @@ public class QueryServiceImpl implements QueryService
 
         if (columnMap.containsKey(fieldKey))
             return null;
+
+        // This could be made more general, but I don't think there's a need.  To reduce testing
+        // just handle expObject() for now, instead of table methods more generally.
+        // Also, we could just expose "objectid" column on all of the experiment tables.
+        if (fieldKey.equals(expObjectIdFieldKey) && table instanceof ExpTable expTable)
+        {
+            ColumnInfo expObjectColumn = expTable.getExpObjectColumn();
+            if (null != expObjectColumn)
+            {
+                ((MutableColumnInfo) expObjectColumn).setFieldKey(expObjectIdFieldKey);
+                columnMap.put(expObjectIdFieldKey, expObjectColumn);
+                return expObjectColumn;
+            }
+            unresolvedColumns.add(new FieldKey(null, "expObject()"));
+            return null;
+        }
 
         ColumnInfo column = getColumn(manager, table, columnMap, fieldKey);
 
@@ -2447,307 +2581,11 @@ public class QueryServiceImpl implements QueryService
     public SQLFragment getSelectSQL(TableInfo table, @Nullable Collection<ColumnInfo> selectColumns, @Nullable Filter filter, @Nullable Sort sort,
                                     int maxRows, long offset, boolean forceSort, @NotNull QueryLogging queryLogging)
     {
-        assert Table.validMaxRows(maxRows) : maxRows + " is an illegal value for rowCount; should be positive, Table.ALL_ROWS or Table.NO_ROWS";
-
-        QueryProfiler.getInstance().ensureListenerEnvironment();
-
-        if (null == selectColumns)
-            selectColumns = table.getColumns();
-        Set<ColumnInfo> allInvolvedColumns = table.getAllInvolvedColumns(selectColumns);
-
-        // Check incoming columns to ensure they come from table
-        assert Table.checkAllColumns(table, selectColumns, "getSelectSQL() selectColumns", true);
-
-        // Create a default sort before ensuring required columns
-        // Don't add a sort if we're running a custom query and it has its own ORDER BY clause
-        boolean viewHasSort = sort != null && !sort.getSortList().isEmpty();
-        boolean viewHasLimit = maxRows > 0 || offset > 0 || Table.NO_ROWS == maxRows;
-        boolean queryHasSort = table instanceof QueryTableInfo && ((QueryTableInfo) table).hasSort();
-        SqlDialect dialect = table.getSqlDialect();
-
-        if (!viewHasSort)
-        {
-            if ((viewHasLimit || forceSort) && (!queryHasSort || dialect.isSqlServer()))
-            {
-                sort = createDefaultSort(selectColumns);
-            }
-        }
-
-        Map<String, SQLFragment> joins = new LinkedHashMap<>();
-        List<ColumnInfo> allColumns = new ArrayList<>(selectColumns);
-        Map<FieldKey, ColumnInfo> columnMap = new HashMap<>();
-        allColumns = ensureRequiredColumns(table, allColumns, filter, sort, null, columnMap, allInvolvedColumns);
-
-        // Check allInvolved columns for which need to be logged
-        // Logged columns may also require data logging (e.g. a patientId)
-        // If a data logging column cannot be found, we will only disallow this query (throw an exception)
-        //      if the logged column that requires data logging is in allColumns (which is selected columns plus ones needed for sort/filter)
-        Map<ColumnInfo, Set<FieldKey>> shouldLogNameToDataLoggingMap = new HashMap<>();
-        Set<ColumnLogging> shouldLogNameLoggings = new HashSet<>();
-        String columnLoggingComment = null;
-        SelectQueryAuditProvider selectQueryAuditProvider = null;
-        for (ColumnInfo column : allInvolvedColumns)
-        {
-            if (!(column instanceof LookupColumn))
-            {
-                ColumnLogging columnLogging = column.getColumnLogging();
-                if (columnLogging.shouldLogName())
-                {
-                    shouldLogNameLoggings.add(columnLogging);
-                    if (!shouldLogNameToDataLoggingMap.containsKey(column))
-                        shouldLogNameToDataLoggingMap.put(column, new HashSet<>());
-                    shouldLogNameToDataLoggingMap.get(column).addAll(columnLogging.getDataLoggingColumns());
-                    if (null == columnLoggingComment)
-                        columnLoggingComment = columnLogging.getLoggingComment();
-                    if (null == selectQueryAuditProvider)
-                        selectQueryAuditProvider = columnLogging.getSelectQueryAuditProvider();
-                }
-            }
-        }
-
-        Set<ColumnInfo> dataLoggingColumns = new HashSet<>();
-        Set<ColumnInfo> extraSelectDataLoggingColumns = new HashSet<>();
-        for (Entry<ColumnInfo, Set<FieldKey>> shouldLogNameToDataLoggingMapEntry : shouldLogNameToDataLoggingMap.entrySet())
-        {
-            for (FieldKey fieldKey : shouldLogNameToDataLoggingMapEntry.getValue())
-            {
-                ColumnInfo loggingColumn = columnMap.get(fieldKey);                 // Look in columnMap
-                if (null == loggingColumn)
-                    loggingColumn = getColumnForDataLogging(table, fieldKey);       // Look in table columns
-
-                if (null == loggingColumn)
-                {
-                    AliasManager manager = new AliasManager(table, allColumns);     // Try to construct column for fieldKey
-                    loggingColumn = getColumn(manager, table, columnMap, fieldKey);
-                }
-
-                if (null != loggingColumn)
-                {
-                    // For the case where we had to add the MRN column in Visualization, that column is in the table.columnMap, but not the local columnMap.
-                    // This is because it's marked as hidden in the queryDef, so not to display, but isn't a normal "extra" column that DataRegion deems to add.
-                    if (!allColumns.contains(loggingColumn))
-                    {
-                        allColumns.add(loggingColumn);
-                        extraSelectDataLoggingColumns.add(loggingColumn);
-                    }
-                    if (!columnMap.containsKey(fieldKey))
-                        columnMap.put(fieldKey, loggingColumn);
-                    dataLoggingColumns.add(loggingColumn);
-                }
-                else
-                {
-                    // Looking for matching column in allColumns; must match by ColumnLogging object, which gets propagated up sql parse tree
-                    for (ColumnInfo column : allColumns)
-                        if (shouldLogNameToDataLoggingMapEntry.getKey().getColumnLogging().equals(column.getColumnLogging()))
-                            throw new UnauthorizedException("Unable to locate required logging column '" + fieldKey.toString() + "'.");
-                }
-            }
-        }
-
-        if (null != table.getUserSchema() && !queryLogging.isReadOnly())
-            queryLogging.setQueryLogging(table.getUserSchema().getUser(), table.getUserSchema().getContainer(), columnLoggingComment,
-                    shouldLogNameLoggings, dataLoggingColumns, selectQueryAuditProvider);
-        else if (!shouldLogNameLoggings.isEmpty())
-            throw new UnauthorizedException("Column logging is required but cannot set query logging object.");
-
-        // Check columns again: ensureRequiredColumns() may have added new columns
-        assert Table.checkAllColumns(table, allColumns, "getSelectSQL() results of ensureRequiredColumns()", true);
-
-        // I think this is for some custom filter/sorts that do not declare fields, but assume all table fields are available
-        for (ColumnInfo c : table.getColumns())
-        {
-            if (!columnMap.containsKey(c.getFieldKey()))
-                columnMap.put(c.getFieldKey(), c);
-        }
-
-        boolean requiresExtraColumns = allColumns.size() > selectColumns.size();
-        SQLFragment outerSelect = new SQLFragment("SELECT *");
-        SQLFragment selectFrag = new SQLFragment("SELECT ");   // SAS/SHARE JDBC driver requires "SELECT " ("SELECT\n" is not allowed), #17168
-        String strComma = "\n";
-        String tableName = table.getName();
-
-        if (tableName == null)
-        {
-            // This shouldn't happen, but if it's null we'll blow up later without enough context to give a good error
-            // message
-            throw new NullPointerException("Null table name from " + table);
-        }
-
-        String tableAlias = AliasManager.makeLegalName(tableName, table.getSchema().getSqlDialect());
-
-        if (allColumns.isEmpty())
-        {
-            selectFrag.append("* ");
-        }
-        else
-        {
-            CaseInsensitiveHashMap<ColumnInfo> aliases = new CaseInsensitiveHashMap<>();
-            ColumnInfo prev;
-            for (ColumnInfo column : allColumns)
-            {
-                if (null != (prev = aliases.put(column.getAlias(), column)))
-                {
-                    if (prev != column)
-                        ExceptionUtil.logExceptionToMothership(null, new Exception("Duplicate alias in column list: " + table.getSchema() + "." + table.getName() + "." + column.getFieldKey().toSQLString() + " as " + column.getAlias()));
-                    continue;
-                }
-                column.declareJoins(tableAlias, joins);
-                selectFrag.append(strComma);
-                selectFrag.append(column.getValueSql(tableAlias));
-                selectFrag.append(" AS ");
-                selectFrag.append(dialect.makeLegalIdentifier(column.getAlias()));
-                strComma = ",\n";
-            }
-        }
-
-        if (requiresExtraColumns)
-        {
-            outerSelect = new SQLFragment("SELECT ");
-            strComma = "";
-
-            for (ColumnInfo column : selectColumns)
-            {
-                outerSelect.append(strComma);
-                outerSelect.append(dialect.getColumnSelectName(column.getAlias()));
-                strComma = ", ";
-            }
-            for (ColumnInfo column : extraSelectDataLoggingColumns)
-            {
-                outerSelect.append(strComma);
-                outerSelect.append(dialect.getColumnSelectName(column.getAlias()));
-                strComma = ", ";
-            }
-        }
-
-        SQLFragment fromFrag = new SQLFragment("FROM ");
-        Set<FieldKey> fieldKeySet = new TreeSet<>();
-        allColumns.stream()
-                .map(col -> col instanceof WrappedColumn ? ((WrappedColumn) col).getWrappedColumn() : col)
-                .forEach(col -> {
-                    var fk = col.getFieldKey();
-                    fieldKeySet.add(fk);
-                    if (null != fk.getParent())
-                        fieldKeySet.add(fk.getRootFieldKey());
-                });
-        SQLFragment getfromsql = table.getFromSQL(tableAlias, fieldKeySet);
-        fromFrag.append(getfromsql);
-        fromFrag.append(" ");
-
-        for (Entry<String, SQLFragment> entry : joins.entrySet())
-        {
-            fromFrag.append("\n").append(entry.getValue());
-        }
-
-        SQLFragment filterFrag = null;
-
-        if (filter != null)
-        {
-            filterFrag = filter.getSQLFragment(dialect, "x", columnMap);
-        }
-
-        String orderBy = null;
-
-        if (sort != null)
-        {
-            orderBy = sort.getOrderByClause(dialect, columnMap);
-        }
-
-        if ((filterFrag == null || filterFrag.getSQL().length() == 0) && sort == null && Table.ALL_ROWS == maxRows && offset == 0)
-        {
-            selectFrag.append("\n").append(fromFrag);
-            return selectFrag;
-        }
-
-        SQLFragment nestedFrom = new SQLFragment();
-        nestedFrom.append("FROM (\n").append(selectFrag).append("\n").append(fromFrag).append(") x");
-        SQLFragment ret = dialect.limitRows(outerSelect, nestedFrom, filterFrag, orderBy, null, maxRows, offset);
-
-        if (AppProps.getInstance().isDevMode())
-        {
-            SQLFragment t = new SQLFragment();
-            t.appendComment("<QueryServiceImpl.getSelectSQL(" + AliasManager.makeLegalName(table.getName(), dialect) + ")>", dialect);
-            t.append(ret);
-            t.appendComment("</QueryServiceImpl.getSelectSQL()>", dialect);
-            ret = SQLFragment.prettyPrint(t);
-        }
-
-        return ret;
+        var query = new Query(table.getUserSchema());
+        var selectView = QuerySelectView.create(query, table, selectColumns, filter, sort, maxRows, offset, forceSort, queryLogging);
+        return selectView.getSql();
     }
 
-
-    private static Sort createDefaultSort(Collection<ColumnInfo> columns)
-    {
-        Sort sort = new Sort();
-        addSortableColumns(sort, columns, true);
-
-        if (sort.getSortList().size() == 0)
-        {
-            addSortableColumns(sort, columns, false);
-        }
-
-        return sort;
-    }
-
-
-    /* see if columns in column.getSortFieldKeys() are resolvable, so we avoid proposing a default sort that won't work */
-    private static List<ColumnInfo> resolveSortFieldKeys(ColumnInfo col, Map<FieldKey, ColumnInfo> selectColumns)
-    {
-        List<FieldKey> sortFieldKeys = col.getSortFieldKeys();
-        if (null != sortFieldKeys && !sortFieldKeys.isEmpty())
-        {
-            // fast way: see if all columns are already selected
-            if (sortFieldKeys.stream().allMatch(f -> null != selectColumns.get(f)))
-                return sortFieldKeys.stream().map(selectColumns::get).collect(Collectors.toList());
-            // slow way: use QueryService
-            Map<FieldKey, ColumnInfo> columns = QueryService.get().getColumns(col.getParentTable(), sortFieldKeys);
-            if (columns.size() == sortFieldKeys.size() && !columns.containsValue(null))
-                return sortFieldKeys.stream().map(f -> columns.get(f)).collect(Collectors.toList());
-        }
-
-        if (!col.isSortable())
-            return null;
-        return Collections.singletonList(col);
-    }
-
-
-    private static void addSortableColumns(Sort sort, Collection<ColumnInfo> columns, boolean usePrimaryKey)
-    {
-	    /* There is a bit of a chicken-and-egg problem here
-	        we need to know what we want to sort on before calling ensureRequiredColumns, but we don't know for sure we
-	        which columns we can sort on until we validate which columns are available (because of getSortFieldKeys)
-	     */
-        Map<FieldKey, ColumnInfo> available = new HashMap<>();
-	    columns.forEach(c -> {if (!available.containsKey(c.getFieldKey())) available.put(c.getFieldKey(),c);});
-
-        for (ColumnInfo column : columns)
-        {
-            if (usePrimaryKey && !column.isKeyField())
-                continue;
-            List<ColumnInfo> sortFields = resolveSortFieldKeys(column, available);
-            if (sortFields != null && !sortFields.isEmpty())
-            {
-                // NOTE: we don't need to expando the list here, Sort.getOrderByClause() will do that
-                sort.appendSortColumn(column.getFieldKey(), column.getSortDirection(), false);
-                return;
-            }
-        }
-    }
-
-    @Nullable
-    private static ColumnInfo getColumnForDataLogging(TableInfo table, FieldKey fieldKey)
-    {
-        ColumnInfo loggingColumn = table.getColumn(fieldKey);
-        if (null != loggingColumn)
-            return loggingColumn;
-
-        // Column names may be mangled by visualization; lookup by original column name
-        for (ColumnInfo column : table.getColumns())
-        {
-            if (column.getColumnLogging().getOriginalColumnFieldKey().equals(fieldKey))
-                return column;
-        }
-        return null;
-    }
 
     @Override
     public void addCompareType(CompareType type)
